update the code structure of global_localizer
